requirements.txt
requests
selenium
pytest
allure-pytest